[TASK] Refactor to support new style importing of apps.
